+++ src/db/index.js
@@ -1,5 +1,5 @@
 const mongoose = require('mongoose');
-const {dbConnectionURL} = require('./config');
+const {dbConnectionURL} = require('../config');
 async function main() {
     await mongoose.connect(dbConnectionURL, { autoIndex: false });
 }
